+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - feat: send telegram notification when a new non scheduled episode is found
 - feat: skip some sources using the environment variable `DISABLED_SOURCES` (`DISABLED_SOURCES=source1,source2`)
 - feat: add command `update-single-source` to update a single source assuming all the episodes are new
+- feat: add command `print` to print the sources definitions and source names
 
 ## [1.0.0] - 2023-03-05